Multiple updates include:
* Adding support for Visual Studio Test Explorer via Google Test Adapter so the tests can be run/debugged in VS
* Fixing multiple warnings - some have major impact on Windows platform/compiler which doesn't automatically zero structs
* Fixing the flaky and explicitly broken unit tests

+++ src/utils/src/ExponentialBackoffRetryStrategy.c
@@ -23,7 +23,7 @@ STATUS normalizeExponentialBackoffConfig(PExponentialBackoffRetryStrategyConfig 
           pExponentialBackoffRetryStrategyConfig->jitterType, pExponentialBackoffRetryStrategyConfig->jitterFactor);
 
     // Seed rand to generate random number for jitter
-    SRAND(GETTIME());
+    SRAND((UINT32) GETTIME());
 
 CleanUp:
     LEAVES();
